# dev/alphacar/backend/api-spec.yaml
openapi: 3.0.3
info:
  title: AlphaCar Backend API
  version: "1.0.0"
  description: |
    AlphaCar 백엔드 통합 API 명세서입니다.
    - 각 서비스는 서로 다른 포트/베이스 URL에서 동작합니다.
    - 아래 paths는 **서비스 기준 상대 경로**로 정의되어 있습니다.
servers:
  - url: http://localhost:3002
    description: Main Service (메인/차량/찜/최근 본 차량/판매)
  - url: http://localhost:3003/api
    description: Quote Service (견적/비교)
  - url: http://localhost:3006
    description: Mypage Service (마이페이지/인증)
  - url: http://localhost:3005
    description: Community/Search Service
  - url: http://localhost:4000
    description: AI Chat Service
  - url: http://localhost:3002
    description: Drive Service (동일 포트 사용 시 게이트웨이 기준으로 조정)
...
paths:
...
  /models:
    get:
      tags: [Main]
      summary: 모델 목록 조회
      parameters:
        - name: makerId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 모델 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /trims:
    get:
      tags: [Main]
      summary: 트림 목록 조회
      parameters:
        - name: modelId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 트림 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /vehicles/{id}:
    get:
      tags: [Vehicles]
      summary: 단일 차량 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 차량 상세 데이터
          content:
            application/json:
              schema:
                type: object
...
  /vehicles/detail:
    get:
      tags: [Quote]
      summary: 견적용 트림 상세 조회
      parameters:
        - name: trimId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 트림 상세
          content:
            application/json:
              schema:
                type: object
  /vehicles/compare-data:
    get:
      tags: [Quote]
      summary: 비교 견적 데이터 조회
      parameters:
        - name: ids
          in: query
          required: true
          description: 콤마로 구분된 트림 ID 목록
          schema:
            type: string
      responses:
        "200":
          description: 비교 데이터 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /auth/kakao-login:
    post:
      tags: [Auth]
      summary: 카카오 로그인 콜백 처리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        "200":
          description: 로그인 결과
          content:
            application/json:
              schema:
                type: object
  /auth/google-login:
    post:
      tags: [Auth]
      summary: 구글 로그인 처리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        "200":
          description: 로그인 결과
          content:
            application/json:
              schema:
                type: object
...
  /chat/ask:
    post:
      tags: [Chat]
      summary: 텍스트 질문 챗봇
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        "200":
          description: 답변
          content:
            application/json:
              schema:
                type: object
  /chat/image:
    post:
      tags: [Chat]
      summary: 이미지 업로드 및 분석
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 분석 결과
          content:
            application/json:
              schema:
                type: object
        "422":
          description: 파일 형식/용량 오류
  /chat/knowledge:
    post:
      tags: [Chat]
      summary: 지식 추가
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                source:
                  type: string
      responses:
        "200":
          description: 추가 결과
          content:
            application/json:
              schema:
                type: object

  ##################################################
  # 9. Drive 서비스 (포트 3002 or 별도)
  ##################################################
...